Modify output, add TestOutputToClient
