+++ main/client/flagsClient/test.txt
@@ -1,0 +1,1 @@
+The server sends: Including key and value in database...
