+++ main/server/entities/command.go
@@ -1,0 +1,1 @@
+package entities
