+++ main/server/flagsServer/validation.go
@@ -1,0 +1,1 @@
+package flagsServer
